refactor(store): Move machine/store to own top level package
The store can be used for other things other than the machines,
so move it to its own top level package.
