+++ store/embedded.go
